Renaming to avoid confusion, sorting files, and better audio
I had named a number of classes Psych*, which led to confusion about whether they belonged to Psychtoolbox
or not. To remedy, I renamed a number of them to Blam*.

Additionally, the desired audio channel for a given input can now be specified, though it doesn't handle
erroneously attempting a stereo input to a mono output.